+++ xivo_confd/config.py
@@ -131,6 +131,7 @@ DEFAULT_CONFIG = {
         'schedule': True,
         'sccp_general': True,
         'sip_general': True,
+        'sound': True,
         'sound_language': True,
         'switchboard': True,
         'switchboard_member': True,